Refactor(*): update components files structure
