+++ client/src/components/footer/footer.js
@@ -1,40 +1,12 @@
 import {
   Box,
-  chakra,
   Container,
   Stack,
-  Text,
   useColorModeValue,
-  VisuallyHidden,
 } from '@chakra-ui/react';
 import { FaLinkedin, FaGithub, FaGoogle } from 'react-icons/fa';
+import { SocialButton } from './social-buttons';
 
-const SocialButton = ({
-  children,
-  label,
-  href,
-}) => {
-  return (
-    <chakra.button
-      bg={useColorModeValue('blackAlpha.100', 'whiteAlpha.100')}
-      rounded={'full'}
-      w={8}
-      h={8}
-      cursor={'pointer'}
-      as={'a'}
-      href={href}
-      display={'inline-flex'}
-      alignItems={'center'}
-      justifyContent={'center'}
-      transition={'background 0.3s ease'}
-      _hover={{
-        bg: useColorModeValue('blackAlpha.200', 'whiteAlpha.200'),
-      }}>
-      <VisuallyHidden>{label}</VisuallyHidden>
-      {children}
-    </chakra.button>
-  );
-};
 
 export default function Footer() {
   return (
@@ -43,13 +15,12 @@ export default function Footer() {
       <Container
         as={Stack}
         maxW={'6xl'}
-        pb={2}
+        bottom={0}
         pt={20}
         direction={{ base: 'column', md: 'row' }}
         spacing={4}
-        justify={{ base: 'center', md: 'space-between' }}
+        justify={{ base: 'center', md: 'center' }}
         align={{ base: 'center', md: 'center' }}>
-        <Text>©2022 All rights reserved</Text>
         <Stack direction={'row'} spacing={6}>
           <SocialButton label={'Github'} href={'https://github.com/Jood80'}>
             <FaGithub />
