--- client/src/components/tweets.js
@@ -1,147 +1,0 @@
-import {
-  Avatar,
-  Box,
-  chakra,
-  Flex,
-  Button,
-  SimpleGrid,
-  useColorModeValue,
-} from '@chakra-ui/react';
-import { AddIcon } from '@chakra-ui/icons';
-import { testimonials } from '../data/dummy-users';
-import { backgrounds } from '../images/backgrounds'
-import { image } from '../images/backgroundImage';
-
-
-function TestmonialCard(props) {
-  const { name, role, content, avatar, index } = props;
-  return (
-    <Flex
-      boxShadow={'lg'}
-      maxW={'640px'}
-      direction={{ base: 'column-reverse', md: 'row' }}
-      width={'full'}
-      rounded={'xl'}
-      p={10}
-      justifyContent={'space-between'}
-      position={'relative'}
-      bg={useColorModeValue('white', 'gray.800')}
-      _after={{
-        content: '""',
-        position: 'absolute',
-        height: '21px',
-        width: '29px',
-        left: '35px',
-        top: '-10px',
-        backgroundSize: 'cover',
-        backgroundImage: image,
-      }}
-      _before={{
-        content: '""',
-        position: 'absolute',
-        zIndex: '-1',
-        height: 'full',
-        maxW: '640px',
-        width: 'full',
-        filter: 'blur(40px)',
-        transform: 'scale(0.98)',
-        backgroundRepeat: 'no-repeat',
-        backgroundSize: 'cover',
-        top: 0,
-        left: 0,
-        backgroundImage: backgrounds[index % 4],
-      }}>
-      <Flex
-        direction={'column'}
-        textAlign={'left'}
-        justifyContent={'space-between'}>
-        <chakra.p
-          fontFamily={'Inter'}
-          fontWeight={'medium'}
-          fontSize={'15px'}
-          pb={4}>
-          {content}
-        </chakra.p>
-        <chakra.p fontFamily={'Work Sans'} fontWeight={'bold'} fontSize={14}>
-          {name}
-          <chakra.span
-            fontFamily={'Inter'}
-            fontWeight={'medium'}
-            color={'gray.500'}>
-            {' '}
-            - {role}
-          </chakra.span>
-        </chakra.p>
-      </Flex>
-      <Avatar
-        src={avatar}
-        height={'80px'}
-        width={'80px'}
-        alignSelf={'center'}
-        m={{ base: '0 0 35px 0', md: '0 0 0 50px' }}
-      />
-    </Flex>
-  );
-}
-
-export default function Tweets() {
-  return (
-    <Flex
-      textAlign={'center'}
-      pt={10}
-      justifyContent={'center'}
-      direction={'column'}
-      width={'full'}>
-      <Box width={{ base: 'full', sm: 'lg', lg: 'xl' }} margin={'auto'}>
-        <chakra.h3
-          fontFamily={'Work Sans'}
-          fontWeight={'bold'}
-          fontSize={20}
-          textTransform={'uppercase'}
-          color={'purple.400'}>
-          People love us
-        </chakra.h3>
-        <chakra.h1
-          py={5}
-          fontSize={48}
-          fontFamily={'Work Sans'}
-          fontWeight={'bold'}
-          color={useColorModeValue('gray.700', 'gray.50')}>
-          You're in good company
-        </chakra.h1>
-        <chakra.h2
-          margin={'auto'}
-          width={'70%'}
-          fontFamily={'Inter'}
-          fontWeight={'medium'}
-          color={useColorModeValue('gray.500', 'gray.400')}>
-          See why over{' '}
-          <chakra.strong color={useColorModeValue('gray.700', 'gray.50')}>
-            150,000+
-          </chakra.strong>{' '}
-          influencers use EEZY to manage their social media content!
-        </chakra.h2>
-        <Button
-          variant={'solid'}
-          colorScheme={'teal'}
-          size={'sm'}
-          mr={4}
-          mt={4}
-          leftIcon={<AddIcon />}>
-          Create Tweet
-        </Button>
-      </Box>
-      <SimpleGrid
-        columns={{ base: 1, xl: 2 }}
-        spacing={'20'}
-        mt={16}
-        mx={'auto'}>
-        {testimonials.map((cardInfo, index) => (
-          <TestmonialCard {...cardInfo} index={index} />
-        ))}
-      </SimpleGrid>
-      <Box>
-      </Box>
-    </Flex>
-  );
-}
