+++ client/src/components/tweets/tweet-card.js
@@ -1,0 +1,82 @@
+import {
+  Avatar,
+  chakra,
+  Flex,
+  useColorModeValue,
+} from '@chakra-ui/react';
+import { backgrounds } from '../../images/backgrounds'
+import { image } from '../../images/backgroundImage';
+
+
+export function TweetCard({ cardInfo, id }) {
+  const { username, tweets, avatar } = cardInfo;
+  const content = tweets[0].content
+
+  return (
+    <Flex
+      boxShadow={'lg'}
+      maxW={'640px'}
+      direction={{ base: 'column-reverse', md: 'row' }}
+      width={'full'}
+      rounded={'xl'}
+      p={10}
+      justifyContent={'space-between'}
+      position={'relative'}
+      bg={useColorModeValue('white', 'gray.800')}
+      _after={{
+        content: '""',
+        position: 'absolute',
+        height: '21px',
+        width: '29px',
+        left: '35px',
+        top: '-10px',
+        backgroundSize: 'cover',
+        backgroundImage: image,
+      }}
+      _before={{
+        content: '""',
+        position: 'absolute',
+        zIndex: '-1',
+        height: 'full',
+        maxW: '640px',
+        width: 'full',
+        filter: 'blur(40px)',
+        transform: 'scale(0.98)',
+        backgroundRepeat: 'no-repeat',
+        backgroundSize: 'cover',
+        top: 0,
+        left: 0,
+        backgroundImage: backgrounds[id % 4],
+      }}>
+      <Flex
+        direction={'column'}
+        textAlign={'left'}
+        justifyContent={'space-between'}>
+        <chakra.p
+          fontFamily={'Inter'}
+          fontWeight={'medium'}
+          fontSize={'15px'}
+          pb={4}>
+          {content}
+        </chakra.p>
+        <chakra.p fontFamily={'Work Sans'} fontWeight={'bold'} fontSize={14}>
+          {username}
+          <chakra.span
+            fontFamily={'Inter'}
+            fontWeight={'medium'}
+            color={'gray.500'}>
+            {' '}
+            - {'role'}
+          </chakra.span>
+        </chakra.p>
+      </Flex>
+      <Avatar
+        src={avatar}
+        height={'80px'}
+        width={'80px'}
+        alignSelf={'center'}
+        m={{ base: '0 0 35px 0', md: '0 0 0 50px' }}
+      />
+    </Flex>
+  );
+}
